requirements.txt
# Core Deep Learning Framework
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
pytorch-lightning==2.3.3
torchmetrics==1.4.1

# Computer Vision & Image Processing
opencv-python-headless==4.10.0.84
pillow==10.4.0
albumentations==2.0.8
scikit-image==0.25.2

# Model Architecture & Pretrained Models
timm==0.9.2
segmentation-models-pytorch==0.3.3
efficientnet-pytorch==0.7.1

# Data Handling & Utilities
numpy==2.2.6
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.0
h5py==3.12.1

# Configuration & Logging
PyYAML==6.0.2
omegaconf==2.3.0
tensorboard==2.18.0
tqdm==4.67.1

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==6.0.1

# Model Analysis & Profiling
thop==0.1.1.post2209072238

# Development & Testing
pytest
jupyter
ipython==8.25.0

# Optional: GPU Support (uncomment if using CUDA)
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cuda-runtime-cu12==12.6.77